refactor: rename temporary components
